removing functionality added + seperated into more classes for structure
