+++ SmartSaver/Form1.cs
@@ -8,7 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace SmartSaver
+namespace DataManager
 {
     public partial class Form1 : Form
     {
